+++ Source/Stubbing/StubFunction/StubThrowingFunction.swift
@@ -9,19 +9,6 @@
 public protocol StubThrowingFunction: StubFunctionThenTrait, StubFunctionThenReturnTrait, StubFunctionThenThrowTrait, StubFunctionThenThrowingTrait {
 }
 
-public protocol StubFunctionThenThrowingTrait: BaseStubFunctionTrait {
-    /// Invokes throwing `implementation` when invoked.
-    func then(_ implementation: @escaping (InputType) throws -> OutputType) -> Self
-}
-
-public extension StubFunctionThenThrowingTrait {
-    @discardableResult
-    func then(_ implementation: @escaping (InputType) throws -> OutputType) -> Self {
-        stub.appendAction(.callImplementation(implementation))
-        return self
-    }
-}
-
 public struct ProtocolStubThrowingFunction<IN, OUT>: StubThrowingFunction {
     public let stub: ConcreteStub<IN, OUT>
